requirements.txt
openai==0.27.7
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.23.1

